src/backend/requirements.txt
psycopg2-binary==2.9.5
pydantic==1.10.5
python-dotenv==1.0.0
SQLAlchemy==1.4.41
asyncpg==0.27.0
fastapi==0.87.0
uvicorn==0.20.0
numpy==1.24.2
opencv-python
face-recognition==1.3.0
SQLAlchemy-ImageAttach==1.1.0
PyJWT==2.6.0 
python-decouple==3.7
bcrypt==4.0.1
python-multipart==0.0.6